Add prod categories type

+++ backend/routes/categories.js
@@ -6,9 +6,19 @@ const checkAuth = require("../middleware/check-auth");
 const router = express.Router();
 
 router.post("", checkAuth, (req, res, next) => {
+    const { name, type } = req.body;
+
+    const categoryType = {
+        typeName: type.typeName,
+        materials: type.materials,
+        shapes: type.shapes,
+        extras: type.extras
+    };
+
     const category = new Category({
-        name: req.body.name,
-        type: [req.body.type]
+        name: name,
+        type: [categoryType],
+        creator: req.userData.userId
     });
     category.save().then( crearedCategory => {
         res.status(201).json({
@@ -19,14 +29,20 @@ router.post("", checkAuth, (req, res, next) => {
 });
 
 router.put("/:id", checkAuth, (req, res, next) => {
-    const category = new Category({
-        name: req.body.name,
-        type: req.body.type
-    });
-    category.updateOne({ _id: req.params.id }, category).then( result => {
+    const categoryType = {
+        typeName: req.body.typeName,
+        materials: req.body.materials,
+        shapes: req.body.shapes,
+        extras: req.body.extras
+    };
+
+    Category.findOneAndUpdate({ _id: req.params.id }, { $push: { type: categoryType } }).then( result => {
         res.status(200).json({
-            message:"Update succesfully"
+            message:"Category update succesfully",
+            result: result
         });
+    }).catch(error => {
+        res.status(500).json({ message: "Couldn't update category" });
     });
 });
 
@@ -49,9 +65,8 @@ router.get("/:id",(req, res, next) => {
 });
 
 router.delete("/:id", checkAuth, (req, res, next) => {
-    Category.deleteOne({_id: req.params.id}).then(result => {
-        console.log(result);
-        res.status(200).json({ message: "Category deleted!"})
+    Category.deleteOne({_id: req.params.id, creator: req.userData.userId}).then(result => {
+        res.status(200).json({ message: "Category deleted!" })
     });
 });
